other pullRequest changes: renaming packages,alligning code styles to the best practices
