+++ README.md
@@ -1,8 +1,7 @@
-
 <h1 align="center">Welcome to final project in yandex automation 4th sprint 👋</h1>
 
 > Author: Mikhail Nersesov
-> 
+>
 > Service under test: https://qa-scooter.praktikum-services.ru/
 
 ## ✨ Technologies in the project
@@ -11,12 +10,9 @@
 >
 > JUnit: 4.13.2
 >
-> Maven: 3.9.2 
-
-
+> Maven: 3.9.2
 
 ## 🚀 Usage
-
 
 To run the tests:
