+++ src/main/java/ru/praktikum/services/qa/scooter/BasePage.java
@@ -1,4 +1,4 @@
-package ru.praktikum_services.qa_scooter;
+package ru.praktikum.services.qa.scooter;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.WebDriverWait;
@@ -6,8 +6,11 @@ import org.openqa.selenium.support.ui.WebDriverWait;
 public abstract class BasePage {
     protected final WebDriver webDriver;
     protected final WebDriverWait webDriverWait;
-    protected  BasePage (WebDriver webDriver){
+
+    protected BasePage(WebDriver webDriver) {
         this.webDriver = webDriver;
-        this.webDriverWait = new WebDriverWait(webDriver,3);
+        this.webDriverWait = new WebDriverWait(webDriver, 3);
     }
+
+
 }
